+++ mapchete/cli/default/index.py
@@ -24,10 +24,11 @@ logger = logging.getLogger(__name__)
 
 @click.command(help="Create index of output tiles.")
 @utils.arg_mapchete_files
-@utils.opt_out_dir
+@utils.opt_idx_out_dir
 @utils.opt_geojson
 @utils.opt_gpkg
 @utils.opt_shp
+@utils.opt_vrt
 @utils.opt_txt
 @utils.opt_fieldname
 @utils.opt_basepath
@@ -41,10 +42,11 @@ logger = logging.getLogger(__name__)
 @utils.opt_logfile
 def index(
     mapchete_files,
-    out_dir=None,
+    idx_out_dir=None,
     geojson=False,
     gpkg=False,
     shp=False,
+    vrt=False,
     txt=False,
     fieldname=None,
     basepath=None,
@@ -57,9 +59,10 @@ def index(
     debug=False,
     logfile=None
 ):
-    if not any([geojson, gpkg, shp, txt]):
+    if not any([geojson, gpkg, shp, txt, vrt]):
         raise click.MissingParameter(
-            "At least one of '--geojson', '--gpkg', '--shp', or '--txt' must be provided.",
+            """At least one of '--geojson', '--gpkg', '--shp', '--vrt' or '--txt'"""
+            """must be provided.""",
             param_type="option"
         )
 
@@ -87,14 +90,16 @@ def index(
                         zoom_index_gen(
                             mp=mp,
                             zoom=tile.zoom,
-                            out_dir=out_dir if out_dir else mp.config.output.path,
+                            out_dir=idx_out_dir if idx_out_dir else mp.config.output.path,
                             geojson=geojson,
                             gpkg=gpkg,
                             shapefile=shp,
+                            vrt=vrt,
                             txt=txt,
                             fieldname=fieldname,
                             basepath=basepath,
-                            for_gdal=for_gdal),
+                            for_gdal=for_gdal
+                        ),
                         total=mp.count_tiles(tile.zoom, tile.zoom),
                         unit="tile",
                         disable=debug
@@ -118,10 +123,13 @@ def index(
                             zoom_index_gen(
                                 mp=mp,
                                 zoom=z,
-                                out_dir=out_dir if out_dir else mp.config.output.path,
+                                out_dir=(
+                                    idx_out_dir if idx_out_dir else mp.config.output.path
+                                ),
                                 geojson=geojson,
                                 gpkg=gpkg,
                                 shapefile=shp,
+                                vrt=vrt,
                                 txt=txt,
                                 fieldname=fieldname,
                                 basepath=basepath,