+++ mapchete/index.py
@@ -8,6 +8,10 @@ Available index types:
 - textfile with tiles list
     If process output is online (e.g. a public endpoint of an S3 container),
     this file can be passed on to wget to download all process output.
+- VRT
+    Virtual raster dataset format by GDAL. This enables GIS tools to read multiple
+    files at once, e.g. QGIS can open a zoom VRT so the user doesn't have to open
+    all GeoTIFF files from a certain zoom level.
 
 All index types are generated once per zoom level. For example GeoPackage will
 generate GPKG files 3.gpkg, 4.gpkg and 5.gpkg for zoom levels 3, 4 and 5.
@@ -19,10 +23,15 @@ from contextlib import ExitStack
 from copy import deepcopy
 import fiona
 import logging
+import operator
 import os
 from shapely.geometry import mapping
+import xml.etree.ElementTree as ET
+from xml.dom import minidom
 
-from mapchete.io import path_exists, get_boto3_bucket
+from mapchete.io import (
+    path_exists, path_is_remote, get_boto3_bucket, raster, relative_path
+)
 
 logger = logging.getLogger(__name__)
 
@@ -45,6 +54,7 @@ def zoom_index_gen(
     gpkg=False,
     shapefile=False,
     txt=False,
+    vrt=False,
     fieldname="location",
     basepath=None,
     for_gdal=True,
@@ -69,6 +79,8 @@ def zoom_index_gen(
         generate Shapefile index (default: False)
     txt : bool
         generate tile path list textfile (default: False)
+    vrt : bool
+        GDAL-style VRT file (default: False)
     fieldname : str
         field name which contains paths of tiles (default: "location")
     basepath : str
@@ -117,6 +129,16 @@ def zoom_index_gen(
             index_writers.append(
                 es.enter_context(
                     TextFileWriter(out_path=_index_file_path(out_dir, zoom, "txt"))
+                )
+            )
+        if vrt:
+            index_writers.append(
+                es.enter_context(
+                    VRTFileWriter(
+                        out_path=_index_file_path(out_dir, zoom, "vrt"),
+                        output=mp.config.output,
+                        out_pyramid=mp.config.output_pyramid
+                    )
                 )
             )
 
@@ -176,7 +198,7 @@ def _tile_path(orig_path, basepath, for_gdal):
 
 
 class VectorFileWriter():
-    """Base class for GeoJSONWriter and GeoPackageWriter."""
+    """Writes GeoJSON or GeoPackage files."""
 
     def __init__(
         self, out_path=None, crs=None, fieldname=None, driver=None
@@ -195,9 +217,9 @@ class VectorFileWriter():
                 logger.debug("read existing entries")
                 with fiona.open(self.path, "r") as src:
                     self._existing = {f["properties"]["tile_id"]: f for f in src}
-                self.file_obj = fiona.open(self.path, "a")
+                self.sink = fiona.open(self.path, "a")
             else:
-                self.file_obj = fiona.open(
+                self.sink = fiona.open(
                     self.path, "w", driver=self.driver, crs=crs, schema=schema
                 )
                 self._existing = {}
@@ -209,10 +231,10 @@ class VectorFileWriter():
                 fiona.remove(self.path, driver=driver)
             else:
                 self._existing = {}
-            self.file_obj = fiona.open(
+            self.sink = fiona.open(
                 self.path, "w", driver=self.driver, crs=crs, schema=schema
             )
-            self.file_obj.writerecords(self._existing.values())
+            self.sink.writerecords(self._existing.values())
 
     def __repr__(self):
         return "VectorFileWriter(%s)" % self.path
@@ -226,7 +248,7 @@ class VectorFileWriter():
     def write(self, tile, path):
         if not self.entry_exists(tile=tile):
             logger.debug("write %s to %s", path, self)
-            self.file_obj.write({
+            self.sink.write({
                 "geometry": mapping(tile.bbox),
                 "properties": {
                     "tile_id": str(tile.id),
@@ -245,7 +267,7 @@ class VectorFileWriter():
 
     def close(self):
         logger.debug("%s new entries in %s", self.new_entries, self)
-        self.file_obj.close()
+        self.sink.close()
 
 
 class TextFileWriter():
@@ -272,9 +294,9 @@ class TextFileWriter():
             self._existing = {}
         self.new_entries = 0
         if self._bucket:
-            self.file_obj = ""
+            self.sink = ""
         else:
-            self.file_obj = open(self.path, "w")
+            self.sink = open(self.path, "w")
         for l in self._existing:
             self._write_line(l)
 
@@ -289,9 +311,9 @@ class TextFileWriter():
 
     def _write_line(self, line):
         if self._bucket:
-            self.file_obj += line
+            self.sink += line
         else:
-            self.file_obj.write(line)
+            self.sink.write(line)
 
     def write(self, tile, path):
         if not self.entry_exists(path=path):
@@ -309,6 +331,171 @@ class TextFileWriter():
         if self._bucket:
             key = "/".join(self.path.split("/")[3:])
             logger.debug("upload %s", key)
-            self.bucket_resource.put_object(Key=key, Body=self.file_obj)
+            self.bucket_resource.put_object(Key=key, Body=self.sink)
         else:
-            self.file_obj.close()
+            self.sink.close()
+
+
+class VRTFileWriter():
+    """Generates GDAL-style VRT file."""
+    def __init__(self, out_path=None, output=None, out_pyramid=None):
+        # see if lxml is installed before checking all output tiles
+        from lxml.builder import ElementMaker
+        self.path = out_path
+        self._tp = out_pyramid
+        self._output = output
+        self._bucket = self.path.split("/")[2] if self.path.startswith("s3://") else None
+        self.bucket_resource = get_boto3_bucket(self._bucket) if self._bucket else None
+        logger.debug("initialize VRT writer for %s", self.path)
+        if path_exists(self.path):
+            if self._bucket:
+                key = "/".join(self.path.split("/")[3:])
+                for obj in self.bucket_resource.objects.filter(Prefix=key):
+                    if obj.key == key:
+                        self._existing = {
+                            k: v for k, v in self._xml_to_entries(
+                                obj.get()['Body'].read().decode()
+                            )
+                        }
+            else:
+                with open(self.path) as src:
+                    self._existing = {k: v for k, v in self._xml_to_entries(src.read())}
+        else:
+            self._existing = {}
+        logger.debug("%s existing entries", len(self._existing))
+        self.new_entries = 0
+        self._new = {}
+
+    def __repr__(self):
+        return "VRTFileWriter(%s)" % self.path
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, *args):
+        self.close()
+
+    def _path_to_tile(self, path):
+        return self._tp.tile(*map(int, os.path.splitext(path)[0].split("/")[-3:]))
+
+    def _add_entry(self, tile=None, path=None):
+        self._new[tile] = path
+
+    def _xml_to_entries(self, xml_string):
+        for entry in next(
+            ET.ElementTree(ET.fromstring(xml_string)).getroot().iter("VRTRasterBand")
+        ).iter("ComplexSource"):
+            path = next(entry.iter("SourceFilename")).text
+            yield (self._path_to_tile(path), path)
+
+    def write(self, tile, path):
+        if not self.entry_exists(tile=tile, path=path):
+            logger.debug("write %s to %s", path, self)
+            self._add_entry(tile=tile, path=path)
+            self.new_entries += 1
+
+    def entry_exists(self, tile=None, path=None):
+        path = relative_path(path=path, base_dir=os.path.split(self.path)[0])
+        exists = path in self._existing
+        logger.debug("tile %s with path %s exists: %s", tile, path, exists)
+        return exists
+
+    def close(self):
+        from lxml.builder import ElementMaker
+
+        logger.debug("%s new entries in %s", self.new_entries, self)
+        if not self._new:
+            logger.debug("no entries to write")
+            return
+
+        # combine existing and new entries
+        all_entries = {**self._existing, **self._new}
+        logger.debug("writing a total of %s entries", len(all_entries))
+
+        # get VRT attributes
+        vrt_affine, vrt_shape = raster.tiles_to_affine_shape(list(all_entries.keys()))
+        vrt_dtype = self._output.profile()["dtype"]
+        vrt_nodata = self._output.nodata
+
+        # build XML
+        E = ElementMaker()
+        vrt = E.VRTDataset(
+            E.SRS(self._tp.crs.to_string()),
+            E.GeoTransform(", ".join(map(str, vrt_affine.to_gdal()))),
+            *[
+                E.VRTRasterBand(
+                    E.NoDataValue(str(vrt_nodata)),
+                    E.ColorInterp("Gray"),
+                    *[
+                        E.ComplexSource(
+                            E.SourceFilename(
+                                relative_path(
+                                    path=path, base_dir=os.path.split(self.path)[0]
+                                ),
+                                relativeToVRT="0" if path_is_remote(path) else "1"
+                            ),
+                            E.SourceBand(str(b_idx)),
+                            E.SourceProperties(
+                                RasterXSize=str(tile.shape.width),
+                                RasterYSize=str(tile.shape.height),
+                                DataType=vrt_dtype,
+                                BlockXSize=str(
+                                    self._output.profile().get(
+                                        "blockxsize", self._tp.tile_size
+                                    )
+                                ),
+                                BlockYSize=str(
+                                    self._output.profile().get(
+                                        "blockysize", self._tp.tile_size
+                                    )
+                                ),
+                            ),
+                            E.SrcRect(
+                                xOff="0",
+                                yOff="0",
+                                xSize=str(tile.shape.width),
+                                ySize=str(tile.shape.height),
+                            ),
+                            E.DstRect(
+                                xOff=str(
+                                    list(raster.bounds_to_ranges(
+                                        out_bounds=tile.bounds,
+                                        in_affine=vrt_affine,
+                                        in_shape=vrt_shape
+                                    ))[2]
+                                ),
+                                yOff=str(
+                                    list(raster.bounds_to_ranges(
+                                        out_bounds=tile.bounds,
+                                        in_affine=vrt_affine,
+                                        in_shape=vrt_shape
+                                    ))[0]
+                                ),
+                                xSize=str(tile.shape.width),
+                                ySize=str(tile.shape.height),
+                            ),
+                            NODATA=str(vrt_nodata)
+                        )
+                        for tile, path in sorted(
+                            all_entries.items(), key=operator.itemgetter(1)
+                        )
+                    ],
+                    dataType=vrt_dtype,
+                    band=str(b_idx)
+                )
+                for b_idx in range(1, self._output.profile()["count"] + 1)
+            ],
+            rasterXSize=str(vrt_shape.width),
+            rasterYSize=str(vrt_shape.height),
+        )
+
+        # generate pretty XML and write
+        xmlstr = minidom.parseString(ET.tostring(vrt)).toprettyxml(indent="  ")
+        if self._bucket:
+            key = "/".join(self.path.split("/")[3:])
+            logger.debug("upload %s", key)
+            self.bucket_resource.put_object(Key=key, Body=xmlstr)
+        else:
+            logger.debug("write to %s", self.path)
+            with open(self.path, "w") as dst:
+                dst.write(xmlstr)